+++ buildchain/buildchain/salt_tree.py
@@ -641,7 +641,6 @@ SALT_FILES: Tuple[Union[Path, targets.AtomicTarget], ...] = (
     Path("salt/_states/metalk8s_sysctl.py"),
     Path("salt/_states/metalk8s_volumes.py"),
     Path("salt/_utils/metalk8s_utils.py"),
-    Path("salt/_utils/kubernetes_utils.py"),
     Path("salt/_utils/pillar_utils.py"),
     Path("salt/_utils/volume_utils.py"),
     # This image is defined here and not in the `image` module since it is